+++ benchmarks/verus/apps/files/apps_test_1017.rs
@@ -18,12 +18,12 @@ spec fn max_distributions(n: int) -> int
 // </vc-helpers>
 
 // <vc-spec>
-fn solve(n: int) -> (result: int)
+fn solve(n: i8) -> (result: i8)
     requires 
-        valid_input(n)
+        valid_input(n as int)
     ensures 
         result >= 1,
-        result == max_distributions(n)
+        result as int == max_distributions(n as int)
 // </vc-spec>
 // <vc-code>
 {
