+++ benchmarks/verus/apps/files/apps_test_4586.rs
@@ -24,11 +24,11 @@ spec fn is_good(n: int) -> bool {
 // </vc-helpers>
 
 // <vc-spec>
-fn solve(n: int) -> (result: bool)
+fn solve(n: i8) -> (result: bool)
     requires 
-        valid_input(n),
+        valid_input(n as int),
     ensures 
-        result <==> is_good(n),
+        result <==> is_good(n as int),
 // </vc-spec>
 // <vc-code>
 {
