+++ benchmarks/verus/numpy_triple/files/indexing_slicing_flatiter.rs
@@ -8,7 +8,7 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn numpy_flatiter(a: Vec<f32>) -> (result: Vec<f32>)
+fn numpy_flatiter(a: Vec<i8>) -> (result: Vec<i8>)
     ensures
         result.len() == a.len(),
         forall|i: int| 0 <= i < result.len() ==> result[i] == a[i],
