+++ benchmarks/verus/numpy_triple/files/strings_isdigit.rs
@@ -13,7 +13,7 @@ spec fn all_chars_digit(s: Seq<char>) -> bool;
 fn isdigit(a: Vec<String>) -> (result: Vec<bool>)
     ensures
         result.len() == a.len(),
-        forall|i: int| 0 <= i < a.len() ==> result[i] == (a[i]@.len() > 0 && all_chars_digit(a[i]@))
+        forall|i: int| 0 <= i < a.len() ==> result[i as int] == (a[i as int]@.len() > 0 && all_chars_digit(a[i as int]@))
 // </vc-spec>
 // <vc-code>
 {
